GH-1748 Renaming e2e tests for checkout flow
