+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/payment-methods.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { user } from '../../sample-data/big-happy-path';
+import { user } from '../../sample-data/checkout-flow';
 import {
   fillShippingAddress,
   fillPaymentDetails
